+++ src/Parsing/Css.php
@@ -837,26 +837,26 @@ class Css
                     }
                     $val = array_values($val);
                     $this->duplicateBorder($val);
-                    $this->value['margin']['t'] = $this->cssConverter->convertToMM($val[0], 0);
-                    $this->value['margin']['r'] = $this->cssConverter->convertToMM($val[1], 0);
-                    $this->value['margin']['b'] = $this->cssConverter->convertToMM($val[2], 0);
-                    $this->value['margin']['l'] = $this->cssConverter->convertToMM($val[3], 0);
+                    $this->value['margin']['t'] = $this->cssConverter->convertToMM($val[0], $this->getLastHeight());
+                    $this->value['margin']['r'] = $this->cssConverter->convertToMM($val[1], $this->getLastWidth());
+                    $this->value['margin']['b'] = $this->cssConverter->convertToMM($val[2], $this->getLastHeight());
+                    $this->value['margin']['l'] = $this->cssConverter->convertToMM($val[3], $this->getLastWidth());
                     break;
 
                 case 'margin-top':
-                    $this->value['margin']['t'] = $this->cssConverter->convertToMM($val, 0);
+                    $this->value['margin']['t'] = $this->cssConverter->convertToMM($val, $this->getLastHeight());
                     break;
 
                 case 'margin-right':
-                    $this->value['margin']['r'] = $this->cssConverter->convertToMM($val, 0);
+                    $this->value['margin']['r'] = $this->cssConverter->convertToMM($val, $this->getLastWidth());
                     break;
 
                 case 'margin-bottom':
-                    $this->value['margin']['b'] = $this->cssConverter->convertToMM($val, 0);
+                    $this->value['margin']['b'] = $this->cssConverter->convertToMM($val, $this->getLastHeight());
                     break;
 
                 case 'margin-left':
-                    $this->value['margin']['l'] = $this->cssConverter->convertToMM($val, 0);
+                    $this->value['margin']['l'] = $this->cssConverter->convertToMM($val, $this->getLastWidth());
                     break;
 
                 case 'border':