+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
   * add `cert` tag - thanks to @LittleBigFox
   * change the name of all the examples from `exemple` to `example`
   * fix issue on qrcode value with multi lines - issue #390
+  * fix documentation - thanks to @noxlux and @tpohchai
 
 ## [5.2.0](https://github.com/spipu/html2pdf/compare/v5.1.0...v5.2.0) - 2018-07-31